requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
mysql-connector-python==8.0.26
python-dotenv==0.19.2
pydantic==1.8.2
passlib[bcrypt]==1.7.4